+++ app/src/main/java/joshuatee/wx/settings/SettingsAboutActivity.kt
@@ -44,6 +44,7 @@ class SettingsAboutActivity : BaseActivity() {
     private val faqUrl = "https://gitlab.com/joshua.tee/wxl23/-/tree/master/doc/FAQ.md"
     //private val releaseNotesUrl = "https://gitlab.com/joshua.tee/wx/-/tree/master/doc/ChangeLog_User.md"
     private val releaseNotesUrl = "https://github.com/ELY3M/wX---modded-by-ELY-M/blob/master/README.md"
+    private val urlPrivacyPolicy = "https://raw.githubusercontent.com/ELY3M/wX---modded-by-ELY-M/refs/heads/master/DOC/PRIVACY_POLICY_FOR_WX.txt"    
 	private val releasesPage = "https://github.com/ELY3M/wX---modded-by-ELY-M/releases"
 	private val latestAPK = "https://github.com/ELY3M/wX---modded-by-ELY-M/releases/latest/download/app-release-unsigned-signed.apk"
     private lateinit var box: VBox
@@ -55,7 +56,12 @@ class SettingsAboutActivity : BaseActivity() {
 
     @SuppressLint("MissingSuperCall")
     override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState, R.layout.activity_linear_layout, R.menu.generic_about, false)
+        super.onCreate(
+            savedInstanceState,
+            R.layout.activity_linear_layout,
+            R.menu.generic_about,
+            false
+        )
         box = VBox.fromResource(this)
         setTitle("About wX", "version " + Utility.getVersion(this))
         addCards()
@@ -64,6 +70,7 @@ class SettingsAboutActivity : BaseActivity() {
     private fun addCards() {
         val faqButton = CardText(this, "View FAQ") { Route.web(this, faqUrl) }
         val releaseNotesButton = CardText(this, "View Release Notes") { Route.web(this, releaseNotesUrl) }
+        val privacyPolicyButton = CardText(this, "Privacy Policy") { Route.web(this, urlPrivacyPolicy) }
 		val releasePageButton = CardText(this, "View Release APKs") { Route.web(this, releasesPage) }
 		val latestAPKButton = CardText(this, "Latest Release APK (Direct Download)") { Route.web(this, latestAPK) }
         val developerSettingsCard = CardText(this, "Developer Settings", SettingsDeveloperActivity::class.java)
@@ -74,6 +81,7 @@ class SettingsAboutActivity : BaseActivity() {
         }
         box.addWidget(faqButton)
         box.addWidget(releaseNotesButton)
+	box.addWidget(privacyPolicyButton)
         box.addWidget(releasePageButton)
 		box.addWidget(latestAPKButton)
         box.addWidget(CardText(this, "Celsius to fahrenheit table") { Route.text(this, UtilityMath.celsiusToFahrenheitTable(), "Celsius to Fahrenheit table") })